+++ BAVCL/Core/GPU/KernelModules/KernelDomain.cs
@@ -1,0 +1,11 @@
+namespace BAVCL;
+
+public enum KernelDomain
+{
+	Arithmetic,
+	Structural,
+	Geometry,
+	Statistics,
+	LinearAlgebra,
+	Astrophysics,
+}
